% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentframecpp-package.R
\docType{package}
\name{parentframecpp-package}
\alias{parentframecpp}
\alias{parentframecpp-package}
\title{parentframecpp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
