% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{via_cpp_cleanup_fixed}
\alias{via_cpp_cleanup_fixed}
\title{Test withr::defer() via C++ (fixed using Rf_eval)}
\usage{
via_cpp_cleanup_fixed()
}
\description{
Test withr::defer() via C++ (fixed using Rf_eval)
}
