% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{via_cpp_error}
\alias{via_cpp_error}
\title{Test rlang::abort() error attribution via C++}
\usage{
via_cpp_error(x = -1)
}
\arguments{
\item{x}{A number}
}
\description{
Test rlang::abort() error attribution via C++
}
