% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{via_cpp_error_fixed}
\alias{via_cpp_error_fixed}
\title{Test rlang::abort() error attribution via C++ (fixed using Rf_eval)}
\usage{
via_cpp_error_fixed(x = -1)
}
\arguments{
\item{x}{A number}
}
\description{
Test rlang::abort() error attribution via C++ (fixed using Rf_eval)
}
