% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{via_r_error}
\alias{via_r_error}
\title{Test rlang::abort() error attribution via pure R}
\usage{
via_r_error(x = -1)
}
\arguments{
\item{x}{A number}
}
\description{
Test rlang::abort() error attribution via pure R
}
