% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentframecpp-package.R
\docType{package}
\name{parentframecpp-package}
\alias{parentframecpp}
\alias{parentframecpp-package}
\title{parentframecpp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jennybc/parentframecpp}
  \item Report bugs at \url{https://github.com/jennybc/parentframecpp/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
